+++ robocorp-code/vscode-client/src/common.ts
@@ -1,0 +1,15 @@
+import { Uri, workspace, WorkspaceFolder } from "vscode";
+
+export const debounce = (func, wait) => {
+    let timeout: NodeJS.Timeout;
+
+    return function wrapper(...args) {
+        const later = () => {
+            clearTimeout(timeout);
+            func(...args);
+        };
+
+        clearTimeout(timeout);
+        timeout = setTimeout(later, wait);
+    };
+};